feat(core): make controller startup importable and its auth replaceable
The controller can only be started by this repository's own main package.
Every wiring decision lives in func main() in package main, which nothing
can import, and the authenticator and authorizer are constructed inline and
cannot be substituted.

Move the startup sequence into go/core/pkg/app behind Run(ctx, Options), and
reduce cmd/controller-v2 to a signal handler and a call. Each Options field
is optional and its zero value selects exactly what the controller does
today, so this adds a seam without changing the shipped binary:

  Authenticator, Authorizer  replace the unsecure and noop defaults
  SetupWithManager           register controllers and scheme types
  HTTPMiddleware             wrap the handler ahead of core's own auth
  Routes                     add endpoints alongside core's
  ExtraMigrations            apply further tracks after the built-ins

Three orderings are load-bearing and are commented where they are relied on.
SetupWithManager runs after the manager exists and before it starts, so a
scheme added there precedes any cache; it also runs before Routes, so a
handler built in the callback can still be registered. ExtraMigrations are
appended rather than merged, because the built-in tracks must reach their
final version before tables that may reference them.

The third is why SetupLogger is exported rather than kept inline. A caller
builds its Options before Run, so anything it logs happens before Run
installs the logger -- and controller-runtime discards everything written
through log.Log until the first SetLogger call. A startup error in that
window is lost entirely and the process appears to die in silence, which is
how this surfaced. Calling SetupLogger first fixes it, and calling it twice
is harmless: SetLogger fulfils a promise that can only be fulfilled once, so
the first caller wins and Run's own call does nothing.

The A2A handler is deliberately not among these. Core builds the gateway
itself, with the instance workflow that lets it suspend an instance at the
end of a turn; a caller supplying its own would get a less complete one, and
the service that reaches agent runtimes should not accept whatever handler a
library consumer assembles.

Most of the diff is a move. 262 of the added lines are the former main()
body, unchanged except for four edits: signal handling stays in main,
log.Fatal becomes a returned error, the hardcoded auth pair becomes
opts.resolve(), and group.Wait() is returned rather than fatal-logged.
Reviewers can confirm that by diffing the two function bodies directly.

One behaviour does change as a result. log.Fatal exits the process, so
db.Close() and actors.Close() never ran on a startup failure. Returning the
error runs them.
